requirements.txt
accelerate==1.5.2
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.2.0
annotated-types==0.6.0
anyio==4.6.2
asgiref==3.8.1
asttokens==3.0.0
attrs==24.3.0
backoff==2.2.1
bcrypt==4.3.0
blinker==1.9.0
Bottleneck==1.4.2
Brotli==1.0.9
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.3.2
chroma-hnswlib==0.7.6
chromadb==0.6.3
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.1
cycler==0.12.1
datasets==3.3.2
decorator==5.2.1
Deprecated==1.2.18
dill==0.3.8
distro==1.9.0
durationpy==0.9
einops==0.8.1
executing==2.2.0
fastapi==0.115.11
filelock==3.13.1
Flask==3.1.0
flatbuffers==25.2.10
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.12.0
google-auth==2.38.0
googleapis-common-protos==1.69.2
grpcio==1.71.0
h11==0.14.0
httpcore==1.0.2
httptools==0.6.4
httpx==0.27.0
huggingface_hub==0.29.2
humanfriendly==10.0
idna==3.7
importlib_metadata==8.6.1
importlib_resources==6.5.2
ipython==9.0.2
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.6.1
joblib==1.4.2
jsonpickle==4.0.2
kiwisolver==1.4.8
kubernetes==32.0.1
lxml==5.3.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mmh3==5.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.15
networkx==3.4.2
nltk==3.9.1
numexpr==2.10.1
numpy==2.2.2
oauthlib==3.2.2
onnxruntime==1.21.0
openai==1.60.1
opentelemetry-api==1.31.1
opentelemetry-exporter-otlp-proto-common==1.31.1
opentelemetry-exporter-otlp-proto-grpc==1.31.1
opentelemetry-instrumentation==0.52b1
opentelemetry-instrumentation-asgi==0.52b1
opentelemetry-instrumentation-fastapi==0.52b1
opentelemetry-proto==1.31.1
opentelemetry-sdk==1.31.1
opentelemetry-semantic-conventions==0.52b1
opentelemetry-util-http==0.52b1
orjson==3.10.15
overrides==7.7.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==11.1.0
pip==25.0
posthog==3.21.0
prompt_toolkit==3.0.50
propcache==0.2.0
protobuf==5.29.4
psutil==7.0.0
pure_eval==0.2.3
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.19.1
pyparsing==3.2.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2024.1
pyvis==0.3.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==3.4.1
setuptools==72.1.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
stack-data==0.6.3
starlette==0.46.1
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.6.0
tokenizers==0.21.0
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tqdm==4.67.1
traitlets==5.14.3
transformers==4.49.0
typer==0.15.2
typing_extensions==4.12.2
tzdata==2023.3
urllib3==2.3.0
uvicorn==0.34.0
watchfiles==1.0.4
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
win_inet_pton==1.1.0
wrapt==1.17.2
xxhash==3.5.0
yarl==1.18.0
zipp==3.21.0
streamlit>=1.40.0
openai>=1.0.0
streamlit>=1.40.0
openai>=1.0.0
sentence-transformers>=3.0.0
einops>=0.8.0
nltk>=3.9.0
python-docx>=1.1.0
chromadb>=0.4.0
anthropic
google-genai
pyvis>=0.3.1
matplotlib>=3.7.0
